Resource Creation Endpoints - I
